==> main/03_construct_figuretable_merger.Rmd <==
---
title: "03_construct_figuretable_merger"
author: "Suguru Otani"
date: "2022/7/2"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(magrittr)
library(ggplot2)
library(modelsummary)
options(modelsummary_format_numeric_latex = "plain") # for modelsummary output
```


```{r cars, warning=FALSE}
operator_level_entry_exit_merger_CIY <-
  readRDS(file = "../output/operator_level_entry_exit_merger_CIY.rds")
operator_level_entry_exit_merger_HBdata <-
  readRDS(file = "../output/operator_level_entry_exit_merger_HBdata.rds")
operator_level_entry_exit_merger_IHS <-
  readRDS(file = "../output/operator_level_entry_exit_merger_IHS.rds")
matching_pair_year_CIY <-
  readRDS(file = "../output/matching_pair_year_CIY.rds")
matching_pair_year_IHS <-
  readRDS(file = "../output/matching_pair_year_IHS.rds")
matching_pair_year_HBdata <-
  readRDS(file = "../output/matching_pair_year_HBdata.rds")
HB_data <-
  readRDS(file = "../output/HB_data.rds")
IHS_data <-
  readRDS(file = "../output/IHS_data.rds")
route_demand_data <-
  readRDS(file = "../output/route_demand_data.rds")
```


# Entry, exit, and mergers {.tabset}


## Table 1: CIY {.tabset}

### Merger list {.tabset}

<!-- ##### Market-level -->

<!-- ```{r, echo=FALSE} -->
<!-- operator_level_entry_exit_merger_CIY_for_output <- -->
<!--   operator_level_entry_exit_merger_CIY %>%  -->
<!--   dplyr::filter( -->
<!--     dummy_merged_until_data_period == 1 -->
<!--   ) %>%  -->
<!--   dplyr::arrange(start) %>%  -->
<!--   dplyr::mutate( -->
<!--     merging_firm_color = -->
<!--       ifelse( -->
<!--         is.na(merging_firm) == 1 & -->
<!--           dummy_merged_until_data_period == 0, -->
<!--              operator, -->
<!--              merging_firm) -->
<!--   ) -->
<!-- operator_level_entry_exit_merger_CIY_for_output %>%  -->
<!--   kableExtra::kable() %>%  -->
<!--   kableExtra::kable_styling() -->

<!-- ``` -->

<!-- ##### Industry-level -->

```{r, echo=FALSE}
merger_list <-
  matching_pair_year_CIY  %>% 
  dplyr::mutate(
    ID = dplyr::row_number()
  ) %>% 
  dplyr::select(
    ID,
    seller_name,
    buyer_name,
    end,
    matching_type
  ) %>% 
  dplyr::rename(
    `Seller` = seller_name,
    `Buyer` = buyer_name,
    `Year` = end,
    `Type` = matching_type
  ) %>% 
  dplyr::mutate(
    `Buyer` =
      ifelse(
        `Buyer` == "P&O Containers",
        "P\\&O Containers",
        `Buyer`
      )
  )
merger_list %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()

merger_list %>% 
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = ""
  ) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/merger_list_CIY.tex")
  )
```

#### Summary statistics

```{r, echo=FALSE}
seller_data <-
  matching_pair_year_CIY %>% 
  dplyr::select(
    #seller_name,
    seller_operator_age_normalized,
    seller_cumsum_TEU_normalized
  ) %>% 
  dplyr::rename(
    `Age (Normalized)` = seller_operator_age_normalized,
    `Size TEU (Normalized)` = seller_cumsum_TEU_normalized
  )
buyer_data <-
  matching_pair_year_CIY %>% 
  dplyr::select(
    #buyer_name,
    buyer_operator_age_normalized,
    buyer_cumsum_TEU_normalized
  ) %>% 
  dplyr::rename(
    `Age (Normalized)` = buyer_operator_age_normalized,
    `Size TEU (Normalized)` = buyer_cumsum_TEU_normalized
  ) 
target_date_CIY <-
  rbind(
    seller_data,
    buyer_data
    ) %>% 
  dplyr::mutate(
    data = "CIY (1966-1990)"
  )
f <- 
  (`Age (Normalized)` +
     `Size TEU (Normalized)`) ~ 
  N + (mean + sd + min + max)*Arguments(na.rm = TRUE)
modelsummary::datasummary(
  f,
  data = target_date_CIY
  )
target_date_CIY %>% 
  modelsummary::datasummary(
    f,
    escape = FALSE,
    data = .,
    output = "latex_tabular"
  ) %>% 
  gsub("\\multicolumn.*\\\\", "", .) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/summary_statistics_of_firms_CIY.tex")
  )
```



#### Overall Transition

```{r, echo=FALSE}
operator_level_entry_exit_merger_CIY_for_output <-
  operator_level_entry_exit_merger_CIY %>% 
  dplyr::distinct(
    operator,
    .keep_all = TRUE
  ) %>% 
  dplyr::arrange(start) %>% 
  dplyr::mutate(
    merging_firm_color =
      ifelse(
        is.na(merging_firm) == 1 &
          dummy_merged_until_data_period == 0,
        operator,
        merging_firm
        )
  )
ggplot(operator_level_entry_exit_merger_CIY_for_output, 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment(aes(colour=merging_firm)) +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

operator_level_entry_exit_merger_CIY_for_output %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()

```



## Table 2: IHS Shipdetails {.tabset}

### Merger list

```{r, echo=FALSE}
merger_list <-
  matching_pair_year_IHS %>% 
  dplyr::mutate(
    ID = dplyr::row_number()
  ) %>% 
  dplyr::select(
    ID,
    seller_name,
    buyer_name,
    end,
    matching_type
  ) %>% 
  dplyr::rename(
    `Seller` = seller_name,
    `Buyer` = buyer_name,
    `Year` = end,
    `Type` = matching_type
  ) %>% 
  dplyr::mutate(
    `Seller` =
      ifelse(
        `Seller` == "ROYAL P&O NEDLLOYD NV (KONINKL",
        "ROYAL P\\&O NEDLLOYD NV",
        `Seller`
      )
  ) %>% 
  dplyr::mutate(
    `Seller` =
      ifelse(
        `Seller` == "IMC SHIPPING CO PTE LTD (IMCSC",
        "IMC SHIPPING CO PTE LTD",
        `Seller`
      )
  ) %>% 
  dplyr::mutate(
    `Buyer` =
      ifelse(
        `Buyer` == "IMC SHIPPING CO PTE LTD (IMCSC",
        "IMC SHIPPING CO PTE LTD",
        `Buyer`
      )
  ) %>% 
  dplyr::mutate(
    `Buyer` =
      ifelse(
        `Buyer` == "A P MOLLER"&
          `Seller` == "FARRELL LINES INC",
        "CSAV",
        `Buyer`
      )
  )
merger_list %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()

merger_list %>% 
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = ""
  ) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/merger_list_IHS.tex")
  )
```


#### Summary statistics

```{r, echo=FALSE}
seller_data <-
  matching_pair_year_IHS %>% 
  dplyr::select(
    #seller_name,
    seller_operator_age_normalized,
    seller_cumsum_TEU_normalized
  ) %>% 
  dplyr::rename(
    `Age (Normalized)` = seller_operator_age_normalized,
    `Size TEU (Normalized)` = seller_cumsum_TEU_normalized
  )
buyer_data <-
  matching_pair_year_IHS %>% 
  dplyr::select(
    #buyer_name,
    buyer_operator_age_normalized,
    buyer_cumsum_TEU_normalized
  ) %>% 
  dplyr::rename(
    `Age (Normalized)` = buyer_operator_age_normalized,
    `Size TEU (Normalized)` = buyer_cumsum_TEU_normalized
  )
target_date_IHS <-
  rbind(
    seller_data,
    buyer_data
    ) %>% 
  dplyr::mutate(
    data = "IHS (1991-2005)"
  )
f <- 
  (`Age (Normalized)` +
     `Size TEU (Normalized)`) ~ 
  N + (mean + sd + min + max)*Arguments(na.rm = TRUE)
modelsummary::datasummary(
  f,
  data = target_date_IHS
  )
target_date_IHS %>% 
  modelsummary::datasummary(
    f,
    escape = FALSE,
    data = .,
    output = "latex_tabular"
  ) %>% 
  gsub("\\multicolumn.*\\\\", "", .) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/summary_statistics_of_firms_IHS.tex")
  )
```



#### Overall Transition

```{r, echo=FALSE}
IHS_data_for_output <-
  IHS_data %>% 
  dplyr::group_by(operator) %>% 
  dplyr::summarise(
    start = min(year),
    end = max(year)
  ) %>% 
  dplyr::ungroup() %>%
  dplyr::arrange(start) 

ggplot(IHS_data_for_output, 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

ggplot(IHS_data_for_output %>% 
         dplyr::filter(start <= 1992), 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

ggplot(IHS_data_for_output %>% 
         dplyr::filter(start <= 1995&
                         start >= 1993), 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

ggplot(IHS_data_for_output %>% 
         dplyr::filter(start <= 1998&
                         start >= 1996), 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

ggplot(IHS_data_for_output %>% 
         dplyr::filter(start <= 2001&
                         start >= 1999), 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

ggplot(IHS_data_for_output %>% 
         dplyr::filter(start <= 2004&
                         start >= 2002), 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

ggplot(IHS_data_for_output %>% 
         dplyr::filter(start <= 2006&
                         start >= 2005), 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator, -start), 
           yend = reorder(operator, -start)
           )) +
  theme_classic() +
  geom_segment() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("operator")

IHS_data_for_output %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()

```



## Table 3: HB_data {.tabset}

### Merger list {.tabset}

#### Merger list (consistent)

```{r, echo=FALSE}
merger_list <-
  matching_pair_year_HBdata %>% 
  dplyr::filter(
    matching_type != "deleted"
  ) %>% 
  dplyr::filter(
    inconsistent_indicator == 0
  ) %>% 
  dplyr::mutate(
    ID = dplyr::row_number()
  ) %>% 
  dplyr::select(
    ID,
    seller_name,
    buyer_name,
    end,
    matching_type
  ) %>% 
  dplyr::rename(
    `Seller` = seller_name,
    `Buyer` = buyer_name,
    `Year` = end,
    `Type` = matching_type
  ) %>% 
  dplyr::mutate(
    ID = dplyr::row_number()
  ) %>% 
  dplyr::mutate(
    `Buyer` =
      ifelse(
        `Buyer` == "CCL",
        "GOTO Shipping International Ltd",
        `Buyer`
      )
  ) 

merger_list %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()

merger_list %>% 
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = ""
  ) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/merger_list_HB.tex")
  )
```

#### Merger list (inconsistent)

```{r, echo=FALSE}
merger_list <-
  matching_pair_year_HBdata %>% 
  dplyr::filter(
    inconsistent_indicator == 1
  ) %>% 
  dplyr::mutate(
    ID = dplyr::row_number()
  ) %>% 
  dplyr::select(
    ID,
    seller_name,
    buyer_name,
    end
  ) %>% 
  dplyr::rename(
    `Seller` = seller_name,
    `Buyer` = buyer_name,
    `Year` = end
  ) %>% 
  dplyr::mutate(
    ID = dplyr::row_number()
  ) %>% 
  dplyr::arrange(`Year`)

merger_list$`Note` <-
  c(
    "Merger occurred in 1999",
    "Merger occurred in 2005",
    "Merger occurred in 1998"
  )
  
merger_list %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()

merger_list %>% 
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = ""
  ) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/merger_list_HB_inconsistent.tex")
  )
```


#### Summary statistics

```{r, echo=FALSE}
seller_data <-
  matching_pair_year_HBdata %>% 
  dplyr::select(
    #seller_name,
    seller_operator_age_normalized,
    seller_cumsum_TEU_normalized
  ) %>% 
  dplyr::rename(
    `Age (Normalized)` = seller_operator_age_normalized,
    `Size TEU (Normalized)` = seller_cumsum_TEU_normalized
  )
buyer_data <-
  matching_pair_year_HBdata %>% 
  dplyr::select(
    #buyer_name,
    buyer_operator_age_normalized,
    buyer_cumsum_TEU_normalized
  ) %>% 
  dplyr::rename(
    `Age (Normalized)` = buyer_operator_age_normalized,
    `Size TEU (Normalized)` = buyer_cumsum_TEU_normalized
  )
target_date_HBdata <-
  rbind(
    seller_data,
    buyer_data
    ) %>% 
  dplyr::mutate(
    data = "HB (2006-2022)"
  )
f <- 
  (`Age (Normalized)` +
     `Size TEU (Normalized)`) ~ 
  N + (mean + sd + min + max)*Arguments(na.rm = TRUE)
modelsummary::datasummary(
  f,
  data = target_date_HBdata
  )
target_date_HBdata %>% 
  modelsummary::datasummary(
    f,
    escape = FALSE,
    data = .,
    output = "latex_tabular"
  ) %>% 
  gsub("\\multicolumn.*\\\\", "", .) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/summary_statistics_of_firms_HB.tex")
  )
```






#### Overall Transition

```{r, echo=FALSE}
operator_level_entry_exit_merger_HBdata_for_output <-
  operator_level_entry_exit_merger_HBdata %>% 
  dplyr::mutate(
    merging_firm_color =
      ifelse(is.na(merging_firm) == 1,
             operator_name,
             merging_firm)
  )

ggplot(operator_level_entry_exit_merger_HBdata_for_output, 
       aes(x = start,
           xend = end + 1, 
           y = reorder(operator_name, -start), 
           yend = reorder(operator_name, -start)
           )) +
  geom_segment(aes(colour=merging_firm)) +
  theme_classic() +
  #theme(legend.position = "none") +
  theme(axis.text.y=element_text(size=rel(0.2))) +
  xlab("Year") +
  ylab("Operator name")
operator_level_entry_exit_merger_HBdata_for_output %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()
```


## Figure 1: The number of mergers between 1966 and 2022

```{r, echo=FALSE}
matching_data <-
  rbind(
    matching_pair_year_CIY,
    matching_pair_year_IHS,
    matching_pair_year_HBdata %>% 
      dplyr::select(- inconsistent_indicator)
  ) %>% 
  dplyr::select(
    seller_name,
    buyer_name,
    end
  ) %>% 
  dplyr::rename(
    `Year` = end
  ) %>% 
  dplyr::group_by(
    `Year`
  ) %>% 
  dplyr::summarise(
    `Number of mergers` =
      dplyr::n()
  ) %>% 
  dplyr::ungroup()
merger_date_for_plot <-
  seq(1966, 2022, 1) %>% 
  tibble::as_tibble() %>% 
  dplyr::rename(
    Year = value
  ) %>% 
  dplyr::left_join(
    matching_data,
    by = c("Year" = "Year")
  ) %>% 
  dplyr::mutate(
    `Number of mergers` =
      ifelse(
        is.na(`Number of mergers`) == 1,
        0,
        `Number of mergers`
      )
  )

x <-
  ggplot(
    merger_date_for_plot,
    aes(
      x = `Year`,
      y = `Number of mergers`
      )
    ) + 
  geom_line(size = 1.0) +
  geom_point() +
  theme_classic() + 
  scale_color_viridis_d() +
  xlim(1966, 2022) +
  ylim(0, 8) +
  ylab("Number of mergers") +
  geom_hline(yintercept = 0) +
  geom_vline(xintercept = 1966, color = "red", alpha = 0.4) + 
  geom_vline(xintercept = 1970, linetype = "longdash") + 
  geom_vline(xintercept = 1980, linetype = "longdash") + 
  geom_vline(xintercept = 1990, linetype = "longdash") + 
  geom_vline(xintercept = 1990, color = "red", alpha = 0.4) + 
  geom_vline(xintercept = 2000, linetype = "longdash") + 
  geom_vline(xintercept = 2005, color = "red", alpha = 0.4) +
  geom_vline(xintercept = 2010, linetype = "longdash") + 
  geom_vline(xintercept = 2020, linetype = "longdash") + 
  geom_vline(xintercept = 2022, color = "red", alpha = 0.4) 
x
figure_name <-
  "../figuretable/number_of_mergers.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```


## Figure 2: Trends in route-year-level shipping prices and quantities. {.tabset}


### (a) The trends of the container freight rate and shipping quantity. {.tabset}

```{r,echo=FALSE,results = 'asis'}
x <- ggplot(route_demand_data,
            aes(x = year, y = p_dollars_per_TEU)) + 
  geom_line(aes(color = route), alpha = 0.6, size = 1.0) +
  geom_point(aes(shape = route, color = route)) +
  theme_classic() + 
  scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  #ggtitle("Freight rate (CPI adjusted to 1995)") +
  theme(plot.title = element_text(hjust = 0.5)) +
  xlab("Year") + 
  ylab("Shipping price (dollars per TEU)") +
  labs(colour = "", fill = "", shape = "") +
  #expand_limits(y=0, x=2000) + 
  geom_vline(xintercept = 1966, linetype = "longdash") + 
  geom_vline(xintercept = 1980, linetype = "longdash") + 
  # annotate("text", x = 1980, y = 5500, size = 1,
  #          label = "Sea-Land withdrawal\nof cartels") +
  geom_vline(xintercept = 1984, linetype = "longdash")# + 
  # annotate("text", x = 1984, y = 5000, size = 1,
  #          label = "The Shipping Act of\n1984 in U.S. enacted")
x
figure_name <- "../figuretable/container_freight_rate_each_route.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```


### (b) Route-level shipping quantity trend between 1966 and 2007

```{r,echo=FALSE,results = 'asis'}
x <- ggplot(route_demand_data,
            aes(x = year, y = q_TEU1000)) + 
  geom_line(aes(color = route), alpha = 0.6, size = 1.0) +
  geom_point(aes(shape = route, color = route)) +
  #geom_point(aes(shape = route, color = route)) +
  theme_classic() +
  scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  #ggtitle("Container shipping quantity (1000 TEU)") +
  theme(plot.title = element_text(hjust = 0.5)) +
  xlab("Year") + 
  ylab("Shipping quantity (1000 TEU)") +
  labs(colour = "", fill = "", shape = "") +
  #expand_limits(y=0, x=2000) + 
  geom_vline(xintercept = 1966, linetype = "longdash") + 
  geom_vline(xintercept = 1980, linetype = "longdash") + 
  # annotate("text", x = 1980, y = 7000, size = 1,
  #          label = "Sea-Land withdrawal\nof cartels") +
  geom_vline(xintercept = 1984, linetype = "longdash") #+ 
  # annotate("text", x = 1984, y = 5000, size = 1, 
  #          label = "The Shipping Act of\n1984 in U.S. enacted")
x
figure_name <- 
  "../figuretable/container_shipping_quantity_each_route.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```



## Figure 3: Distributions of match-level distances of seller and buyer firms for each regime {.tabset}

### Size distribution

```{r, echo=FALSE}
size_age_data <-
  rbind(
    target_date_CIY,
    target_date_IHS,
    target_date_HBdata
    )
x <-
  ggplot(
    size_age_data, 
    aes(`Size TEU (Normalized)`,
        shape = data, color = data)
    ) +
  stat_ecdf(geom = "step", alpha = 0.6)  +
  theme_classic() + 
  scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=6)) +
  ylab("CDF") +
  xlim(0, 1)
x
figure_name <- 
  "../figuretable/normalized_size_cdf.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Age distribution

```{r, echo=FALSE}
x <-
  ggplot(
    size_age_data, 
    aes(`Age (Normalized)`,
        shape = data, color = data)
    ) +
  stat_ecdf(geom = "step", alpha = 0.6)  +
  theme_classic() + 
  scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=6)) +
  ylab("CDF") +
  xlim(0, 1)
x
figure_name <- 
  "../figuretable/normalized_age_cdf.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```



## Figure 4: Distance distribution

```{r, echo=FALSE}
distance_data_CIY <-
  matching_pair_year_CIY %>% 
  dplyr::select(
    seller_name,
    buyer_name,
    seller_lat,
    seller_lon,
    buyer_lat,
    buyer_lon,
  ) %>% 
  dplyr::mutate(
    `Distance between firms' countries (1000km)` = 
      acos(
        sin(seller_lat)*sin(buyer_lat) + 
          cos(seller_lat) * cos(buyer_lat) *
          cos(seller_lon - buyer_lon)
        ) * 6378.137/ 1000 
      # sqrt(
      #   abs(seller_lat - buyer_lat)^2 + 
      #     abs(seller_lon - buyer_lon)^2
      # )
  ) %>% 
  dplyr::mutate(
    data = 
      "CIY(1966-1990)"
  )
distance_data_IHS <-
  matching_pair_year_IHS %>% 
  dplyr::select(
    seller_name,
    buyer_name,
    seller_lat,
    seller_lon,
    buyer_lat,
    buyer_lon,
  ) %>% 
  dplyr::mutate(
    `Distance between firms' countries (1000km)` = 
      acos(
        sin(seller_lat)*sin(buyer_lat) + 
          cos(seller_lat) * cos(buyer_lat) *
          cos(seller_lon - buyer_lon)
        ) * 6378.137/ 1000 
  ) %>% 
  dplyr::mutate(
    data = 
      "IHS(1991-2005)"
  )
distance_data_HB <-
  matching_pair_year_HBdata %>% 
  dplyr::select(
    seller_name,
    buyer_name,
    seller_lat,
    seller_lon,
    buyer_lat,
    buyer_lon,
  ) %>% 
  dplyr::mutate(
    `Distance between firms' countries (1000km)` = 
      acos(
        sin(seller_lat)*sin(buyer_lat) + 
          cos(seller_lat) * cos(buyer_lat) *
          cos(seller_lon - buyer_lon)
        ) * 6378.137/ 1000 
  ) %>% 
  dplyr::mutate(
    data = 
      "HB(2006-2022)"
  )

distance_data <-
  rbind(
    distance_data_CIY,
    distance_data_IHS,
    distance_data_HB
  )

x <-
  ggplot(
    distance_data, 
    aes(`Distance between firms' countries (1000km)`,
        shape = data, color = data)
    ) +
  stat_ecdf(geom = "step", alpha = 0.6)  +
  theme_classic() + 
  scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=6)) +
  ylab("CDF") +
  xlim(0, 20.4)
x
figure_name <- 
  "../figuretable/distance_cdf.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```
