==> main/05_2_construct_figuretable_maximum_score.Rmd <==
---
title: "05_2_construct_figuretable_maximum_score.Rmd"
author: "Suguru Otani"
date: "2022/7/2"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list = ls())
library(magrittr)
library(ggplot2)
```


```{r cars, warning=FALSE, echo=FALSE}
matching_pair_year_IHS <-
  readRDS(file = "../output/matching_pair_year_IHS.rds")
matching_pair_year_HBdata <-
  readRDS(file = "../output/matching_pair_year_HBdata.rds")
param_list_CIY <-
  readr::read_csv("../output/param_list_CIY.csv") %>% 
  dplyr::rename(
    beta1 = x1,
    beta2 = x2,
    correct_match_percent = x3
  ) %>% 
  dplyr::filter(
    correct_match_percent >= max(correct_match_percent)
  ) %>% 
  dplyr::mutate(
    min_beta1 = 
      format(round(min(beta1), digits = 3), nsmall = 3),
    max_beta1 =
      format(round(max(beta1), digits = 3), nsmall = 3),
    min_beta2 = 
      format(round(min(beta2), digits = 3), nsmall = 3),
    max_beta2 =
      format(round(max(beta2), digits = 3), nsmall = 3),
    max_correct_match_percent = 
      format(round(max(correct_match_percent), digits = 3), nsmall = 3)
  )
param_list_IHS <-
  readr::read_csv("../output/param_list_IHS.csv") %>% 
  dplyr::rename(
    beta1 = x1,
    beta2 = x2,
    correct_match_percent = x3
  ) %>% 
  dplyr::filter(
    correct_match_percent >= max(correct_match_percent)
  ) %>% 
  dplyr::mutate(
    min_beta1 = 
      format(round(min(beta1), digits = 3), nsmall = 3),
    max_beta1 =
      format(round(max(beta1), digits = 3), nsmall = 3),
    min_beta2 = 
      format(round(min(beta2), digits = 3), nsmall = 3),
    max_beta2 =
      format(round(max(beta2), digits = 3), nsmall = 3),
    max_correct_match_percent = 
      format(round(max(correct_match_percent), digits = 3), nsmall = 3)
  )

param_list_HB <-
  readr::read_csv("../output/param_list_HB.csv") %>% 
  dplyr::rename(
    beta1 = x1,
    beta2 = x2,
    correct_match_percent = x3
  ) %>% 
  dplyr::filter(
    correct_match_percent >= max(correct_match_percent)
  ) %>% 
  dplyr::mutate(
    min_beta1 = 
      format(round(min(beta1), digits = 3), nsmall = 3),
    max_beta1 =
      format(round(max(beta1), digits = 3), nsmall = 3),
    min_beta2 = 
      format(round(min(beta2), digits = 3), nsmall = 3),
    max_beta2 =
      format(round(max(beta2), digits = 3), nsmall = 3),
    max_correct_match_percent = 
      format(round(max(correct_match_percent), digits = 3), nsmall = 3)
  )

counterfactual_IHS_max_beta <-
  readr::read_csv("../output/counterfactual_IHS_max_beta.csv")
counterfactual_IHS_min_beta <-
  readr::read_csv("../output/counterfactual_IHS_min_beta.csv")
counterfactual_HB_max_beta <-
  readr::read_csv("../output/counterfactual_HB_max_beta.csv")
counterfactual_HB_min_beta <-
  readr::read_csv("../output/counterfactual_HB_min_beta.csv")
predicted_IHS_max_beta <-
  readr::read_csv("../output/predicted_IHS_max_beta.csv")
predicted_IHS_min_beta <-
  readr::read_csv("../output/predicted_IHS_min_beta.csv")
predicted_HB_max_beta <-
  readr::read_csv("../output/predicted_HB_max_beta.csv")
predicted_HB_min_beta <-
  readr::read_csv("../output/predicted_HB_min_beta.csv")
# operator_level_entry_exit_merger_IHS <-
#   readr::read_csv("../input/operator_level_entry_exit_merger_IHS.csv") %>% 
#   dplyr::distinct(
#     parent_company,
#     .keep_all = TRUE
#     ) 
generate_counterfactual_to_benchmark <-
  function(
    matching_pair_year_IHS,
    predicted_IHS_max_beta,
    counterfactual_IHS_max_beta,
    matching_pair_year_HBdata,
    predicted_HB_max_beta,
    counterfactual_HB_max_beta,
    target_simulation_id
    ){
    matching_pair_year_IHS_different_country <-
      matching_pair_year_IHS %>% 
      # dplyr::filter(
      #   seller_lat != buyer_lat
      # ) %>% 
      dplyr::mutate(
        matching_lat =
          paste(seller_lat, buyer_lat)
      ) %>% 
      dplyr::select(
        matching_lat
      ) %>% 
      dplyr::arrange(matching_lat)
    matching_pair_year_IHS_predicted <-
      predicted_IHS_max_beta %>% 
      dplyr::mutate(
        matching_lat =
          paste(seller_lat, buyer_lat)
      ) %>%
      dplyr::filter(
        simulation_id == target_simulation_id
      ) %>% 
      dplyr::select(
        matching_lat
      ) %>% 
      dplyr::arrange(matching_lat)
    matching_pair_year_IHS_counterfactual <-
      counterfactual_IHS_max_beta %>% 
      dplyr::mutate(
        matching_lat =
          paste(seller_lat, buyer_lat)#,
        # matching_id =
        #   paste(seller_lat, buyer_lat, seller_operator_age_normalized, buyer_operator_age_normalized)
      ) %>%
      # pick up the most common matching
      # dplyr::group_by(matching_id) %>% 
      # dplyr::mutate(
      #   frequency = dplyr::n()
      # ) %>% 
      # dplyr::ungroup() %>% 
      dplyr::filter(
        simulation_id == target_simulation_id
      ) %>% 
      dplyr::select(
        matching_lat
      ) %>% 
      dplyr::arrange(matching_lat)
    matching_pair_year_HB_different_country <-
      matching_pair_year_HBdata %>% 
      # dplyr::filter(
      #   seller_lat != buyer_lat &
      #     seller_lon != buyer_lon
      # ) %>% 
      dplyr::mutate(
        matching_lat =
          paste(seller_lat, buyer_lat)
      ) %>% 
      dplyr::select(
        matching_lat
      ) %>% 
      dplyr::arrange(matching_lat)
    matching_pair_year_HB_predicted <-
      predicted_HB_max_beta %>% 
      dplyr::mutate(
        matching_lat =
          paste(seller_lat, buyer_lat)
      ) %>% 
      dplyr::filter(
        simulation_id == target_simulation_id
      ) %>% 
      dplyr::select(
        matching_lat
      ) %>% 
      dplyr::arrange(matching_lat)
    matching_pair_year_HB_counterfactual <-
      counterfactual_HB_max_beta %>% 
      dplyr::filter(
        simulation_id == target_simulation_id
      ) %>% 
      dplyr::mutate(
        matching_lat =
          paste(seller_lat, buyer_lat)
      ) %>% 
      dplyr::select(
        matching_lat
      ) %>% 
      dplyr::arrange(matching_lat)
    # probability
    rate_number_of_merger_IHS <-
      length(matching_pair_year_IHS_counterfactual$matching_lat)/
      length(matching_pair_year_IHS_different_country$matching_lat)
    
    rate_number_of_same_merger_IHS <-
      1 - 
      length(
        setdiff(
          matching_pair_year_IHS_counterfactual$matching_lat,
          matching_pair_year_IHS_different_country$matching_lat
        )
      )/length(matching_pair_year_IHS_different_country$matching_lat)
    rate_number_of_merger_HB <-
      length(matching_pair_year_HB_counterfactual$matching_lat)/
      length(matching_pair_year_HB_different_country$matching_lat)
    rate_number_of_same_merger_HB <-
      1 - 
      length(
        setdiff(
          matching_pair_year_HB_counterfactual$matching_lat,
          matching_pair_year_HB_different_country$matching_lat
        )
      )/length(matching_pair_year_HB_predicted$matching_lat)
  res <-
    list(
      rate_number_of_merger_IHS = rate_number_of_merger_IHS,
      rate_number_of_same_merger_IHS = rate_number_of_same_merger_IHS,
      rate_number_of_merger_HB = rate_number_of_merger_HB,
      rate_number_of_same_merger_HB = rate_number_of_same_merger_HB,
      matching_pair_year_IHS_predicted = matching_pair_year_IHS_predicted,
      matching_pair_year_IHS_counterfactual = matching_pair_year_IHS_counterfactual,
      matching_pair_year_HB_predicted = matching_pair_year_HB_predicted,
      matching_pair_year_HB_counterfactual = matching_pair_year_HB_counterfactual
    )
  return(res)
}



```


# Mergers {.tabset}

## Maximum score estimator {.tabset}

```{r, echo=FALSE}

result_for_output <-
  cbind(
    rbind(
      "Firm age: $\\beta_1$",
      "Firm size (TEU): $\\beta_2$",
      "Distance: $\\beta_3$",
      "",
      "\\% of correct matches"
    ),
    rbind(
      "",
      "",
      "",
      "",
      ""
    ),
    rbind(
      "1",
      paste(
        "[",
        param_list_CIY$min_beta1[1] ,",",
        param_list_CIY$max_beta1[1] , "]",
        sep = ""
        ),
      paste(
        "[",
        param_list_CIY$min_beta2[1] ,",",
        param_list_CIY$max_beta2[1] , "]",
        sep = ""
        ),
      "",
      param_list_CIY$max_correct_match_percent[1]
      ),
    # IHS
    rbind(
      "1",
      paste(
        "[",
        param_list_IHS$min_beta1[1] ,",",
        param_list_IHS$max_beta1[1] , "]", 
        sep = ""
        ),
      paste(
        "[",
        param_list_IHS$min_beta2[1] ,",",
        param_list_IHS$max_beta2[1] , "]", 
        sep = ""
        ),
      "",
      param_list_IHS$max_correct_match_percent[1]
      ),
    # HB
    rbind(
      "1",
      paste(
        "[",
        param_list_HB$min_beta1[1] ,",",
        param_list_HB$max_beta1[1] , "]", 
        sep = ""
        ),
      paste(
        "[",
        param_list_HB$min_beta2[1] ,",",
        param_list_HB$max_beta2[1] , "]", 
        sep = ""
        ),
      "",
      param_list_HB$max_correct_match_percent[1]
      )
    )
result_for_output <-
  rbind(
    c("Regime", "", "1966-1990", "1991-2005" , "2006-2022"),
    c("", "", "", "" , ""),
    result_for_output
  )

result_for_output %>%
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = "",
    align = "lcccc"
  ) %>%
  kableExtra::row_spec(1, hline_after = TRUE) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/maximum_score_estimate.tex")
  )
result_for_output %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()
```


## Counterfactual{.tabset}

### Upper bounds of beta

```{r, echo=FALSE}
generate_counterfactual_table <-
  function(
    matching_pair_year_IHS,
    predicted_IHS_max_beta,
    counterfactual_IHS_max_beta,
    matching_pair_year_HBdata,
    predicted_HB_max_beta,
    counterfactual_HB_max_beta
    ){
    rate_number_of_merger_IHS = c(1:100)
    rate_number_of_same_merger_IHS = c(1:100)
    rate_number_of_merger_HB = c(1:100)
    rate_number_of_same_merger_HB = c(1:100)
    matching_pair_year_IHS_predicted = c(1:100)
    matching_pair_year_IHS_counterfactual = c(1:100)
    matching_pair_year_HB_predicted = c(1:100)
    matching_pair_year_HB_counterfactual = c(1:100)
    for(i in 1:100){
      temp_list <-
        generate_counterfactual_to_benchmark(
          matching_pair_year_IHS,
          predicted_IHS_max_beta,
          counterfactual_IHS_max_beta,
          matching_pair_year_HBdata,
          predicted_HB_max_beta,
          counterfactual_HB_max_beta,
          target_simulation_id = i
          )
      rate_number_of_merger_IHS[i] = temp_list$rate_number_of_merger_IHS
      rate_number_of_same_merger_IHS[i] = temp_list$rate_number_of_same_merger_IHS
      rate_number_of_merger_HB[i] = temp_list$rate_number_of_merger_HB
      rate_number_of_same_merger_HB[i] = temp_list$rate_number_of_same_merger_HB
      matching_pair_year_IHS_predicted[i] = dim(temp_list$matching_pair_year_IHS_predicted)[1]
      matching_pair_year_IHS_counterfactual[i] = dim(temp_list$matching_pair_year_IHS_counterfactual)[1]
      matching_pair_year_HB_predicted[i] = dim(temp_list$matching_pair_year_HB_predicted)[1]
      matching_pair_year_HB_counterfactual[i] = dim(temp_list$matching_pair_year_HB_counterfactual)[1]
    }
    rate_number_of_merger_IHS <-
      paste(
        "[",
        format(round(
          min(rate_number_of_merger_IHS),
        digits = 3), nsmall = 3),
        ",",
        format(round(
          max(rate_number_of_merger_IHS),
        digits = 3), nsmall = 3),
        "]",
        sep = ""
      )
    rate_number_of_same_merger_IHS <-
      paste(
        "[",
        format(round(
          min(rate_number_of_same_merger_IHS),
          digits = 3), nsmall = 3),
        ",",
        format(round(
          max(rate_number_of_same_merger_IHS),
          digits = 3), nsmall = 3),
        "]",
        sep = ""
      )
    matching_number_IHS <-
      paste(
        "[",
        format(round(
          min(matching_pair_year_IHS_predicted),
          digits = 3), nsmall = 3),
        ",",
        format(round(
          max(matching_pair_year_IHS_predicted),
          digits = 3), nsmall = 3),
        "]",
        sep = ""
      )
    
    rate_number_of_merger_HB <-
      paste(
        "[",
        format(round(
          min(rate_number_of_merger_HB),
        digits = 3), nsmall = 3),
        ",",
        format(round(
          max(rate_number_of_merger_HB),
        digits = 3), nsmall = 3),
        "]",
        sep = ""
      )
    rate_number_of_same_merger_HB <-
      paste(
        "[",
        format(round(
          min(rate_number_of_same_merger_HB),
          digits = 3), nsmall = 3),
        ",",
        format(round(
          max(rate_number_of_same_merger_HB),
          digits = 3), nsmall = 3),
        "]",
        sep = ""
      )
    matching_number_HB <-
      paste(
        "[",
        format(round(
          min(matching_pair_year_HB_predicted),
          digits = 3), nsmall = 3),
        ",",
        format(round(
          max(matching_pair_year_HB_predicted),
          digits = 3), nsmall = 3),
        "]",
        sep = ""
      )
      
    
    result_for_output <-
      rbind(
        c("Regime", "", "1991-2005" , "2006-2022"),
        c(
          "Matching Num (data)", 
          "",  
          matching_number_IHS,
          matching_number_HB
          ),
        # c("Matching Num (counterfactual)", "",  
        #   length(matching_pair_year_IHS_counterfactual$matching_lat),
        #   length(matching_pair_year_HB_counterfactual$matching_lat)),
        c(
          "\\% total match (counterfactual/data)",
          "",  
          rate_number_of_merger_IHS ,
          rate_number_of_merger_HB
          ),
        #c("","","",""),
        c(
          "\\% same match (counterfactual/data)",
          "",  
          rate_number_of_same_merger_HB, 
          rate_number_of_same_merger_HB
          )
        
      )
    return(result_for_output)
  }

result_for_output <-
  generate_counterfactual_table(
    matching_pair_year_IHS,
    predicted_IHS_max_beta,
    counterfactual_IHS_max_beta,
    matching_pair_year_HBdata,
    predicted_HB_max_beta,
    counterfactual_HB_max_beta
    )
result_for_output %>%
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = "",
    align = "lccc"
  ) %>%
  kableExtra::row_spec(1, hline_after = TRUE) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/number_of_mergers_counterfactual_upper_bound.tex")
  )
result_for_output %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()
```


### Lower bounds of beta

```{r, echo=FALSE}
result_for_output <-
  generate_counterfactual_table(
    matching_pair_year_IHS,
    predicted_IHS_min_beta,
    counterfactual_IHS_min_beta,
    matching_pair_year_HBdata,
    predicted_HB_min_beta,
    counterfactual_HB_min_beta
    )

result_for_output %>%
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = "",
    align = "lccc"
  ) %>%
  kableExtra::row_spec(1, hline_after = TRUE) %>% 
  kableExtra::save_kable(
    file = here::here("figuretable/number_of_mergers_counterfactual_lower_bound.tex")
  )
result_for_output %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()
```
